move configs to config dir
