+++ API/config/inversify.config.ts
@@ -1,5 +1,5 @@
 import {Container} from "inversify";
-import {iocTypes} from "./ioc-types";
+import {iocTypes} from "../ioc-types";
 import {AppConfig} from "./app-config";
 
 export const iocContainer = new Container();
